% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AB}
\alias{AB}
\title{Public policy data on air pollution in the Valdivia commune, Chile.}
\format{
Three-dimensional matrix with 16 incentives in rows, 4 behavior in columns and 10 key experts.

Incentives
\describe{
  \item{I1}{ Prohibition}
  \item{I2}{ Standard(regulation))}
  \item{I3}{ Subsidy for heating devices for public services}
  \item{I4}{ Subsidy for heating devices for households}
  \item{I5}{ Record (inventory, information)}
  \item{I6}{ Dry firewood}
  \item{I7}{ Inspection and audit}
  \item{I8}{ Improvement and production}
  \item{I9}{ Subsidy for firewood}
  \item{I10}{ Subsidy for thermal insulation of houses}
  \item{I11}{ Education}
  \item{I12}{ Subsidy for boilers}
  \item{I13}{ R&D, research and development}
  \item{I14}{ Subsidy for transportation}
  \item{I15}{ Environmental impact assessment system}
  \item{I16}{ Mitigation}
}

Behavior
\describe{
  \item{B1}{ People improve thermal insulation of houses}
  \item{B2}{ People improve firewood combustion equipment efficiency and its components}
  \item{B3}{ People improve firewood quality and availability of other combustibles}
  \item{B4}{ Education and awareness in the community}
}
}
\usage{
AB
}
\description{
We illustrate the use of the <package.name> package with data on public policies
 on air pollution from the Valdivia commune, Chile. Studied by Manna, E. M et al (2018) where
  incentives were considered to promote behavior change in public policy on air pollution.
}
\details{
AB
}
\references{
Nicolas, C., Rojas-Mora, J., & Valenzuela-Fernandez, L. (2020). Causal relationships between economic activity and the mining industry in Chile. Journal of Intelligent & Fuzzy Systems, 38(5), 5405-5412.
}
\keyword{datasets}
