% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootMargin.R
\name{bootMargin}
\alias{bootMargin}
\title{bootMargin}
\usage{
bootMargin(
  CC,
  CE = NULL,
  EE = NULL,
  thr = 0.5,
  reps = 10000,
  conf.level = 0.95,
  delete = FALSE,
  plot = FALSE
)
}
\arguments{
\item{CC}{bootMargin}

\item{CE}{CE}

\item{EE}{EE}

\item{thr}{bootMargin}

\item{reps}{bootMargin}

\item{conf.level}{conf.level}

\item{delete}{bootMargin}

\item{plot}{bootMargin}
}
\value{
bootMargin
}
\description{
bootMargin
}
\examples{
bootMargin(CC = AA, CE = AB, EE = BB, thr = 0.5,
 reps = 1000, conf.level = 0.05,
  delete = FALSE, plot = FALSE)
}
