% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{centrality}
\usage{
centrality(
  CC,
  CE = NULL,
  EE = NULL,
  model = c("conpl", "conlnorm", "median"),
  reps = 10000,
  conf = 0.95,
  parallel = c("multicore", "snow", "no"),
  ncpus = 1
)
}
\arguments{
\item{CC}{centrality}

\item{CE}{ce}

\item{EE}{ee}

\item{model}{centrality}

\item{reps}{centrality}

\item{conf}{centrality}

\item{parallel}{centrality}

\item{ncpus}{centrality}
}
\value{
centrality
}
\description{
centrality
}
\examples{
c1 <- centrality( CC = AA, model = "median", reps = 100, parallel = "no", ncpus = 1)
c2 <- centrality( CC = AA, model = "conpl", reps = 100, parallel = "no", ncpus = 1)
c3 <- centrality( CC = AA, model = "conlnorm", reps = 100, parallel = "no", ncpus = 1)
}
