% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEffects.R
\name{directEffects}
\alias{directEffects}
\title{directEffects}
\usage{
directEffects(
  CC,
  CE = NULL,
  EE = NULL,
  thr = 0.5,
  conf.level = 0.95,
  reps = 10000,
  delete = FALSE
)
}
\arguments{
\item{CC}{directEffects}

\item{CE}{directEffects}

\item{EE}{directEffects}

\item{thr}{directEffects}

\item{conf.level}{directEffects}

\item{reps}{directEffects}

\item{delete}{directEffects}
}
\value{
directEffects
}
\description{
directEffects
}
\examples{
directEffects(CC = AA, CE = AB, EE = BB, thr = 0.5, conf.level=0.95, reps = 100, delete = FALSE)
}
