% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgottenEffects.R
\name{fe.rect}
\alias{fe.rect}
\title{Forgotten Effects For Bipartite Graphs}
\usage{
fe.rect(
  CC,
  CE,
  EE,
  thr = 0.5,
  maxOrder = 2,
  reps = 10000,
  parallel = c("multicore", "snow", "no"),
  ncpus = 1
)
}
\arguments{
\item{CC}{Three-dimensional matrix, where each submatrix along the z-axis is a square and reflective incidence matrix, or a list of data.frames containing square and reflective incidence matrices.}

\item{CE}{Three-dimensional matrix, where each submatrix along the z-axis is a reflective rectangular incidence matrix, or a list of data.frames containing reflective and rectangular incidence matrices.}

\item{EE}{Three-dimensional matrix, where each submatrix along the z-axis is a square and reflective incidence matrix, or a list of data.frames containing square and reflective incidence matrices.}

\item{thr}{Real within [0,1]. Defines the significant degree of truth.}

\item{maxOrder}{Positive integer greater than 1: defines the maximum order of the forgotten effects. By default maxOrder = 2.}

\item{reps}{Number of bootstrap replicas. By default reps = 10,000.}

\item{parallel}{The type of parallel operation that is used (if applicable). The options are "multicore", "snow" and "no". By default parallel = "no"}

\item{ncpus}{Integer, Number of processes that are used in the parallel implementation. By default ncpus = 1.}
}
\value{
asd
}
\description{
Performs the forgotten effects calculation proposed by Kaufman and Gil-Aluja (1988) with multiple experts. The parameters allow you to specify the significant degree of truth and the order of incidence that is required to be calculated for rectangular matrices.
}
\details{
The function extends the theory of forgotten effects proposed by Kaufman and Gil-Aluja (1988), to find indirect cause-effect relationships from direct cause-effect relationships, in the case of multiple experts. Indirect cause-effect relationships are classified according to the order to which they correspond.
The parallel and ncpus options are not available on Windows operating systems.
}
\examples{
fe.rect( CC = AA, CE = AB, EE = BB, thr = 0.5, maxOrder = 2, reps = 100, parallel = "no", ncpus = 1)
}
